refactor: fixing tests, moving utils around, added util for in-line function
